# Refactor prefix and add slog handlers (gelf & logrus text)