--- a/logrus_text_formatter.go
+++ b/logrus_text_formatter.go
@@ -3,18 +3,13 @@ package logger
 import (
 	"bytes"
 	"fmt"
-	"io"
-	"os"
 	"regexp"
 	"sort"
 	"strings"
 	"sync"
 	"sync/atomic"
-	"time"
 
-	"github.com/mgutz/ansi"
 	"github.com/sirupsen/logrus"
-	"golang.org/x/term"
 )
 
 // A LogrusTextFormatter is pretty printer for Logrus.
@@ -91,6 +86,11 @@ func (f *LogrusTextFormatter) SetColorScheme(colorScheme *ColorScheme) {
 
 // Format implements logrus.Formatter.
 func (f *LogrusTextFormatter) Format(entry *logrus.Entry) ([]byte, error) {
+	if prefix, ok := entry.Data[KeyPrefix]; ok {
+		entry.Message = fmt.Sprintf("%s %s", prefix, entry.Message)
+		delete(entry.Data, KeyPrefix)
+	}
+
 	keys := make([]string, 0, len(entry.Data))
 	for k := range entry.Data {
 		keys = append(keys, k)
@@ -266,113 +266,6 @@ func (f *LogrusTextFormatter) init(entry *logrus.Entry) {
 	f.template = " %s=" + f.ValueFormatter
 
 	if entry.Logger != nil {
-		f.isTerminal = f.checkIfTerminal(entry.Logger.Out)
-	}
-}
-
-func (f *LogrusTextFormatter) checkIfTerminal(w io.Writer) bool {
-	switch v := w.(type) {
-	case *os.File:
-		return term.IsTerminal(int(v.Fd()))
-	default:
-		return false
-	}
-}
-
-// This is to not silently overwrite `time`, `msg` and `level` fields when
-// dumping it. If this code wasn't there doing:
-//
-//	logrus.WithField("level", 1).Info("hello")
-//
-// would just silently drop the user provided level. Instead with this code
-// it'll be logged as:
-//
-//	{"level": "info", "fields.level": 1, "msg": "hello", "time": "..."}
-func prefixFieldClashes(data logrus.Fields) {
-	if t, ok := data["time"]; ok {
-		data["fields.time"] = t
-	}
-
-	if m, ok := data["msg"]; ok {
-		data["fields.msg"] = m
-	}
-
-	if l, ok := data["level"]; ok {
-		data["fields.level"] = l
-	}
-}
-
-const defaultTimestampFormat = time.RFC3339
-
-var (
-	baseTimestamp      = time.Now()
-	defaultColorScheme = &ColorScheme{
-		InfoLevelStyle:  "green",
-		WarnLevelStyle:  "yellow",
-		ErrorLevelStyle: "red",
-		FatalLevelStyle: "red",
-		PanicLevelStyle: "red",
-		DebugLevelStyle: "blue",
-		PrefixStyle:     "cyan",
-		TimestampStyle:  "black+h",
-	}
-
-	noColorsColorScheme = &compiledColorScheme{
-		InfoLevelColor:  ansi.ColorFunc(""),
-		WarnLevelColor:  ansi.ColorFunc(""),
-		ErrorLevelColor: ansi.ColorFunc(""),
-		FatalLevelColor: ansi.ColorFunc(""),
-		PanicLevelColor: ansi.ColorFunc(""),
-		DebugLevelColor: ansi.ColorFunc(""),
-		PrefixColor:     ansi.ColorFunc(""),
-		TimestampColor:  ansi.ColorFunc(""),
-	}
-
-	defaultCompiledColorScheme = compileColorScheme(defaultColorScheme)
-)
-
-func miniTS() int {
-	return int(time.Since(baseTimestamp) / time.Second)
-}
-
-type ColorScheme struct {
-	InfoLevelStyle  string
-	WarnLevelStyle  string
-	ErrorLevelStyle string
-	FatalLevelStyle string
-	PanicLevelStyle string
-	DebugLevelStyle string
-	PrefixStyle     string
-	TimestampStyle  string
-}
-
-type compiledColorScheme struct {
-	InfoLevelColor  func(string) string
-	WarnLevelColor  func(string) string
-	ErrorLevelColor func(string) string
-	FatalLevelColor func(string) string
-	PanicLevelColor func(string) string
-	DebugLevelColor func(string) string
-	PrefixColor     func(string) string
-	TimestampColor  func(string) string
-}
-
-func getCompiledColor(main string, fallback string) func(string) string {
-	if main == "" {
-		main = fallback
-	}
-	return ansi.ColorFunc(main)
-}
-
-func compileColorScheme(s *ColorScheme) *compiledColorScheme {
-	return &compiledColorScheme{
-		InfoLevelColor:  getCompiledColor(s.InfoLevelStyle, defaultColorScheme.InfoLevelStyle),
-		WarnLevelColor:  getCompiledColor(s.WarnLevelStyle, defaultColorScheme.WarnLevelStyle),
-		ErrorLevelColor: getCompiledColor(s.ErrorLevelStyle, defaultColorScheme.ErrorLevelStyle),
-		FatalLevelColor: getCompiledColor(s.FatalLevelStyle, defaultColorScheme.FatalLevelStyle),
-		PanicLevelColor: getCompiledColor(s.PanicLevelStyle, defaultColorScheme.PanicLevelStyle),
-		DebugLevelColor: getCompiledColor(s.DebugLevelStyle, defaultColorScheme.DebugLevelStyle),
-		PrefixColor:     getCompiledColor(s.PrefixStyle, defaultColorScheme.PrefixStyle),
-		TimestampColor:  getCompiledColor(s.TimestampStyle, defaultColorScheme.TimestampStyle),
+		f.isTerminal = checkIfTerminal(entry.Logger.Out)
 	}
 }