--- a/null.go
+++ b/null.go
@@ -7,73 +7,73 @@ func NewNullLogger() Logger {
 	return &null{}
 }
 
-func (w *null) WithPrefix(prefix string) Logger {
+func (w *null) WithPrefix(_ string) Logger {
 	return w
 }
 
-func (w *null) WithPrefixf(prefix string, args ...any) Logger {
+func (w *null) WithPrefixf(_ string, _ ...any) Logger {
 	return w
 }
 
-func (w *null) WithField(key string, value any) Logger {
+func (w *null) WithField(_ string, _ any) Logger {
 	return w
 }
 
-func (w *null) WithError(err error) Logger {
+func (w *null) WithError(_ error) Logger {
 	return w
 }
 
-func (w *null) WithFields(fields map[string]any) Logger {
+func (w *null) WithFields(_ map[string]any) Logger {
 	return w
 }
 
-func (w *null) Debug(args ...any) {
+func (w *null) Debug(_ ...any) {
 }
 
-func (w *null) Debugf(format string, args ...any) {
+func (w *null) Debugf(_ string, _ ...any) {
 }
 
-func (w *null) Info(args ...any) {
+func (w *null) Info(_ ...any) {
 }
 
-func (w *null) Infof(format string, args ...any) {
+func (w *null) Infof(_ string, _ ...any) {
 }
 
-func (w *null) Warn(args ...any) {
+func (w *null) Warn(_ ...any) {
 }
 
-func (w *null) Warnf(format string, args ...any) {
+func (w *null) Warnf(_ string, _ ...any) {
 }
 
-func (w *null) Error(args ...any) {
+func (w *null) Error(_ ...any) {
 }
 
-func (w *null) Errorf(format string, args ...any) {
+func (w *null) Errorf(_ string, _ ...any) {
 }
 
-func (w *null) Print(args ...any) {
+func (w *null) Print(_ ...any) {
 }
 
-func (w *null) Printf(format string, args ...any) {
+func (w *null) Printf(_ string, _ ...any) {
 }
 
-func (w *null) Println(args ...any) {
+func (w *null) Println(_ ...any) {
 }
 
-func (w *null) Fatal(args ...any) {
+func (w *null) Fatal(_ ...any) {
 }
 
-func (w *null) Fatalf(format string, args ...any) {
+func (w *null) Fatalf(_ string, _ ...any) {
 }
 
-func (w *null) Fatalln(args ...any) {
+func (w *null) Fatalln(_ ...any) {
 }
 
-func (w *null) Panic(args ...any) {
+func (w *null) Panic(_ ...any) {
 }
 
-func (w *null) Panicf(format string, args ...any) {
+func (w *null) Panicf(_ string, _ ...any) {
 }
 
-func (w *null) Panicln(args ...any) {
+func (w *null) Panicln(_ ...any) {
 }