--- a/slog.go
+++ b/slog.go
@@ -7,21 +7,19 @@ import (
 )
 
 type slogwrapper struct {
-	prefix string
-	slog   *slog.Logger
+	slog *slog.Logger
 }
 
 // WrapSlog returns Logger based on log/slog backend.
 func WrapSlog(l *slog.Logger) Logger {
 	return &slogwrapper{
-		slog: cloneslog(l),
+		slog: l,
 	}
 }
 
 func (w *slogwrapper) WithPrefix(prefix string) Logger {
 	return &slogwrapper{
-		prefix: w.prefix + prefix,
-		slog:   cloneslog(w.slog),
+		slog: w.slog.With(KeyPrefix, prefix),
 	}
 }
 
@@ -31,65 +29,61 @@ func (w *slogwrapper) WithPrefixf(format string, args ...any) Logger {
 
 func (w *slogwrapper) WithField(key string, value any) Logger {
 	return &slogwrapper{
-		prefix: w.prefix,
-		slog:   w.slog.With(key, value),
+		slog: w.slog.With(key, value),
 	}
 }
 
 func (w *slogwrapper) WithError(err error) Logger {
 	return &slogwrapper{
-		prefix: w.prefix,
-		slog:   w.slog.With("error", err),
+		slog: w.slog.With("error", err),
 	}
 }
 
 func (w *slogwrapper) WithFields(fields map[string]any) Logger {
-	l := &slogwrapper{
-		prefix: w.prefix,
-		slog:   cloneslog(w.slog),
-	}
-
+	l := w.slog
 	for k, v := range fields {
-		l.slog = l.slog.With(k, v)
+		l = l.With(k, v)
 	}
 
-	return l
+	return &slogwrapper{
+		slog: l,
+	}
 }
 
 func (w *slogwrapper) Debug(args ...any) {
-	w.slog.Debug(w.args(args))
+	w.slog.Debug(fmt.Sprint(args...))
 }
 
 func (w *slogwrapper) Debugf(format string, args ...any) {
-	w.slog.Debug(w.format(format, args))
+	w.slog.Debug(fmt.Sprintf(format, args...))
 }
 
 func (w *slogwrapper) Info(args ...any) {
-	w.slog.Info(w.args(args))
+	w.slog.Info(fmt.Sprint(args...))
 }
 
 func (w *slogwrapper) Infof(format string, args ...any) {
-	w.slog.Info(w.format(format, args))
+	w.slog.Info(fmt.Sprintf(format, args...))
 }
 
 func (w *slogwrapper) Warn(args ...any) {
-	w.slog.Warn(w.args(args))
+	w.slog.Warn(fmt.Sprint(args...))
 }
 
 func (w *slogwrapper) Warnf(format string, args ...any) {
-	w.slog.Warn(w.format(format, args))
+	w.slog.Warn(fmt.Sprintf(format, args...))
 }
 
 func (w *slogwrapper) Error(args ...any) {
-	w.slog.Error(w.args(args))
+	w.slog.Error(fmt.Sprint(args...))
 }
 
 func (w *slogwrapper) Errorf(format string, args ...any) {
-	w.slog.Error(w.format(format, args))
+	w.slog.Error(fmt.Sprintf(format, args...))
 }
 
 func (w *slogwrapper) Print(args ...any) {
-	w.slog.Info(w.args(args))
+	w.slog.Info(fmt.Sprint(args...))
 }
 
 func (w *slogwrapper) Printf(format string, args ...any) {
@@ -135,39 +129,7 @@ func (w *slogwrapper) Panicln(args ...any) {
 //
 //
 
-func (w *slogwrapper) format(format string, args []any) string {
-	return fmt.Sprintf(w.appendPrefix(format), args...)
-}
-
-func (w *slogwrapper) args(args []any) string {
-	args = w.prependPrefix(args)
-	return fmt.Sprint(args...)
-}
-
 func (w *slogwrapper) argsln(args []any) string {
-	args = w.prependPrefix(args)
 	msg := fmt.Sprintln(args...)
 	return msg[:len(msg)-1]
-}
-
-func (w *slogwrapper) appendPrefix(msg string) string {
-	if w.prefix == "" {
-		return msg
-	}
-
-	return fmt.Sprintf("%s %s", w.prefix, msg)
-}
-
-func (w *slogwrapper) prependPrefix(args []any) []any {
-	if w.prefix == "" {
-		return args
-	}
-
-	return append([]any{w.prefix + " "}, args...)
-}
-
-// aka l.clone()
-func cloneslog(l *slog.Logger) *slog.Logger {
-	c := *l
-	return &c
 }