Resolve #8

Comparison between first implementation and optimised implementation for GELF:
```
[logger]>> benchstat testouille/base.txt testouille/new.txt
goos: darwin
goarch: amd64
pkg: github.com/mdouchement/logger
cpu: Intel(R) Core(TM) i7-4870HQ CPU @ 2.50GHz
             │ testouille/base.txt │         testouille/new.txt          │
             │       sec/op        │   sec/op     vs base                │
LogrusGELF-8           8.992µ ± 1%   6.893µ ± 5%  -23.34% (p=0.000 n=10)
SlogGELF-8             8.309µ ± 1%   5.952µ ± 2%  -28.37% (p=0.000 n=10)
LogrusText-8           8.013µ ± 1%   7.895µ ± 1%   -1.47% (p=0.000 n=10)
SlogText-8             7.353µ ± 0%   7.476µ ± 0%   +1.67% (p=0.000 n=10)
geomean                8.145µ        7.015µ       -13.88%

             │ testouille/base.txt │          testouille/new.txt          │
             │        B/op         │     B/op      vs base                │
LogrusGELF-8          5.933Ki ± 0%   4.482Ki ± 1%  -24.45% (p=0.000 n=10)
SlogGELF-8            5.074Ki ± 0%   2.639Ki ± 0%  -47.99% (p=0.000 n=10)
LogrusText-8          4.519Ki ± 0%   4.513Ki ± 0%        ~ (p=0.084 n=10)
SlogText-8            5.329Ki ± 0%   5.431Ki ± 0%   +1.92% (p=0.000 n=10)
geomean               5.189Ki        4.126Ki       -20.47%

             │ testouille/base.txt │          testouille/new.txt          │
             │      allocs/op      │ allocs/op   vs base                  │
LogrusGELF-8            76.00 ± 0%   65.00 ± 0%  -14.47% (p=0.000 n=10)
SlogGELF-8              85.00 ± 0%   73.00 ± 0%  -14.12% (p=0.000 n=10)
LogrusText-8            65.00 ± 0%   65.00 ± 0%        ~ (p=1.000 n=10) ¹
SlogText-8              72.00 ± 0%   74.00 ± 0%   +2.78% (p=0.000 n=10)
geomean                 74.15        69.12        -6.79%
¹ all samples are equal
```

Comparison between first implementation and #10:
```
[logger]>> benchstat testouille/base.txt testouille/new2.txt
goos: darwin
goarch: amd64
pkg: github.com/mdouchement/logger
cpu: Intel(R) Core(TM) i7-4870HQ CPU @ 2.50GHz
             │ testouille/base.txt │         testouille/new2.txt          │
             │       sec/op        │    sec/op     vs base                │
LogrusGELF-8           8.992µ ± 1%   6.693µ ± 14%  -25.57% (p=0.000 n=10)
SlogGELF-8             8.309µ ± 1%   4.657µ ±  3%  -43.96% (p=0.000 n=10)
LogrusText-8           8.013µ ± 1%   7.638µ ±  0%   -4.68% (p=0.000 n=10)
SlogText-8             7.353µ ± 0%   6.244µ ±  1%  -15.08% (p=0.000 n=10)
geomean                8.145µ        6.209µ        -23.77%

             │ testouille/base.txt │         testouille/new2.txt          │
             │        B/op         │     B/op      vs base                │
LogrusGELF-8          5.933Ki ± 0%   4.487Ki ± 1%  -24.37% (p=0.000 n=10)
SlogGELF-8            5.074Ki ± 0%   2.429Ki ± 0%  -52.14% (p=0.000 n=10)
LogrusText-8          4.519Ki ± 0%   4.504Ki ± 0%   -0.31% (p=0.000 n=10)
SlogText-8            5.329Ki ± 0%   5.571Ki ± 0%   +4.55% (p=0.000 n=10)
geomean               5.189Ki        4.067Ki       -21.63%

             │ testouille/base.txt │         testouille/new2.txt          │
             │      allocs/op      │ allocs/op   vs base                  │
LogrusGELF-8            76.00 ± 0%   65.00 ± 0%  -14.47% (p=0.000 n=10)
SlogGELF-8              85.00 ± 0%   63.00 ± 0%  -25.88% (p=0.000 n=10)
LogrusText-8            65.00 ± 0%   65.00 ± 0%        ~ (p=1.000 n=10) ¹
SlogText-8              72.00 ± 0%   64.00 ± 0%  -11.11% (p=0.000 n=10)
geomean                 74.15        64.24       -13.36%
¹ all samples are equal
```